added folders, exported functions, big clean up of the entire backend
