+++ .gitignore
@@ -3,6 +3,7 @@
 # keys
 serviceAccountKey.json
 oauthKey.js
+keys.js
 
 # dependencies
 node_modules/
