+++ src/backend/routes/menu.js
@@ -1,17 +1,8 @@
 const express = require('express');
 
 const router = express.Router();
-const admin = require('firebase-admin');
 
-const serviceAccount = require('../serviceAccountKey.json');
-
-// initialize an instance of firestore with admin priviledges
-admin.initializeApp({
-  credential: admin.credential.cert(serviceAccount),
-});
-
-// creates the db variable to be accessed
-const db = admin.firestore();
+const db = require('../services/firestore');
 
 // Routes
